requirements.txt
Flask==2.1.0
Pillow==9.0.0
requests==2.26.0
numpy==1.21.0
torch==1.10.0
torchvision==0.11.0
urllib3==1.26.7